Use link instead of button for export
