+++ src/pages/messages/components/NviPublicationPointsPage.tsx
@@ -25,8 +25,9 @@ export const NviPublicationPointsPage = () => {
   return (
     <NviStatusWrapper
       headline={headline}
-      exportAcronym={organizationQuery.data?.acronym}
-      topView={<NviPublicationPointsTexts aggregationsQuery={nviStatusQuery} />}
+      topView={
+        <NviPublicationPointsTexts exportAcronym={organizationQuery.data?.acronym} aggregationsQuery={nviStatusQuery} />
+      }
       visibilitySelector>
       <TableContainer component={Paper} variant="outlined">
         <Table size="small">
